Add support for reporting state
* cannot include the 'success' tag in a state update
  - need to redo a bit around state reporting
* broke out & make possible others to run the example program
