+++ device.go
@@ -1,5 +1,7 @@
 package action
 
+import "encoding/json"
+
 // DeviceName contains different ways of identifying the device
 type DeviceName struct {
 	// DefaultNames (not user settable)
@@ -117,4 +119,157 @@ func NewSwitch(id string) *Device {
 	d := NewDevice(id, "action.devices.types.SWITCH")
 	d.AddOnOffTrait(false, false)
 	return d
+}
+
+// AddBrightnessTrait indicates this device is capable of having its brightness controlled.
+// If the device does not support querying, set onlyCommand to true (i.e. a write-only switch).
+// See https://developers.google.com/assistant/smarthome/traits/brightness
+func (d *Device) AddBrightnessTrait(onlyCommand bool) *Device {
+	d.Traits["action.devices.traits.Brightness"] = true
+	if onlyCommand {
+		d.attributes["commandOnlyBrightness"] = true
+	}
+
+	return d
+}
+
+// ColorModel defines which model of the color wheel the device supports.
+const (
+	RGB = "rgb"
+	HSV = "hsv"
+)
+
+// AddColourTrait indicates this device is capable of having its colour display controlled using the specified color model.
+// It is mutually exclusive to support RGB or HSV.
+// It is possible to support either one of RGB and HSV alongside color temperature. See AddColorTemperatureSetting
+// If the device does not support querying, set onlyCommand to true (i.e. a write-only lightbulb).
+// See https://developers.google.com/assistant/smarthome/traits/colorsetting
+func (d *Device) AddColourTrait(model string, onlyCommand bool) *Device {
+	d.Traits["action.devices.traits.ColorSetting"] = true
+	if onlyCommand {
+		d.attributes["commandOnlyColorSetting"] = true
+	}
+	d.attributes["colorModel"] = model
+
+	return d
+}
+
+// AddColourTemperatureTrait indicates this device is capable of having its colour display controlled using the colour temperature model.
+// This can be set alongside AddColorSetting to indicate both color temperature and another algorithm are supported.
+// If the device does not support querying, set onlyCommand to true (i.e. a write-only lightbulb).
+// See https://developers.google.com/assistant/smarthome/traits/colorsetting
+func (d *Device) AddColourTemperatureTrait(minTempK int, maxTempK int, onlyCommand bool) *Device {
+	d.Traits["action.devices.traits.ColorSetting"] = true
+
+	if onlyCommand {
+		d.attributes["commandOnlyColorSetting"] = true
+	} else {
+		d.attributes["commandOnlyColorSetting"] = false
+	}
+	d.attributes["colorTemperatureRange"] = map[string]int{
+		"temperatureMinK": minTempK,
+		"temperatureMaxK": maxTempK,
+	}
+
+	return d
+}
+
+// AddInputSelectorTrait indicates this device is capable of having its input selected.
+// See https://developers.google.com/assistant/smarthome/traits/inputselector
+func (d *Device) AddInputSelectorTrait(availableInputs []DeviceInput, ordered bool) *Device {
+	d.Traits["action.devices.traits.InputSelector"] = true
+	d.attributes["availableInputs"] = availableInputs
+	d.attributes["orderedInputs"] = ordered
+
+	return d
+}
+
+// AddOnOffTrait indicates this device is capable of having its state toggled on or off.
+// If the device can be commanded but not queried, set onlyCommand to true (i.e. a write-only switch).
+// If the devie cannot be commanded but only queried, set onlyQuery to true (i.e. a sensor).
+// See https://developers.google.com/assistant/smarthome/traits/onoff
+func (d *Device) AddOnOffTrait(onlyCommand, onlyQuery bool) *Device {
+	d.Traits["action.devices.traits.OnOff"] = true
+	if onlyCommand {
+		d.attributes["commandOnlyOnOff"] = true
+	}
+	if onlyQuery {
+		d.attributes["queryOnlyOnOff"] = true
+	}
+
+	return d
+}
+
+// AddVolumeTrait indicates this device is capable of having its volume controlled
+// See https://developers.google.com/assistant/smarthome/traits/volume
+func (d *Device) AddVolumeTrait(maxLevel int, canMute bool, onlyCommand bool) *Device {
+	d.Traits["action.devices.traits.Volume"] = true
+	if onlyCommand {
+		d.attributes["commandOnlyVolume"] = true
+	}
+	d.attributes["volumeMaxLevel"] = maxLevel
+	d.attributes["volumeCanMuteAndUnmute"] = canMute
+
+	return d
+}
+
+// MarshalJSON is a custom JSON serializer for our Device
+func (d Device) MarshalJSON() ([]byte, error) {
+	dr := deviceRaw{}
+
+	dr.ID = d.ID
+	dr.Type = d.Type
+	for trait := range d.Traits {
+		dr.Traits = append(dr.Traits, trait)
+	}
+	dr.Name.DefaultNames = d.Name.DefaultNames
+	dr.Name.Name = d.Name.Name
+	dr.Name.Nicknames = d.Name.Nicknames
+	dr.WillReportState = d.WillReportState
+	dr.RoomHint = d.RoomHint
+	dr.Attributes = d.attributes
+	dr.DeviceInfo.Manufacturer = d.DeviceInfo.Manufacturer
+	dr.DeviceInfo.Model = d.DeviceInfo.Model
+	dr.DeviceInfo.HwVersion = d.DeviceInfo.HwVersion
+	dr.DeviceInfo.SwVersion = d.DeviceInfo.SwVersion
+	for _, otherDeviceID := range d.OtherDeviceIDs {
+		dr.OtherDeviceIDs = append(dr.OtherDeviceIDs, otherDeviceIDraw{
+			AgentID:  otherDeviceID.AgentID,
+			DeviceID: otherDeviceID.DeviceID,
+		})
+	}
+	dr.CustomData = d.CustomData
+
+	return json.Marshal(dr)
+}
+
+type otherDeviceIDraw struct {
+	AgentID  string `json:"agentId,omitempty"`
+	DeviceID string `json:"deviceId,omitempty"`
+}
+
+type deviceRaw struct {
+	ID     string   `json:"id,omitempty"`
+	Type   string   `json:"type,omitempty"`
+	Traits []string `json:"traits,omitempty"`
+
+	Name struct {
+		DefaultNames []string `json:"defaultNames,omitempty"`
+		Name         string   `json:"name,omitempty"`
+		Nicknames    []string `json:"nicknames,omitempty"`
+	} `json:"name,omitempty"`
+
+	WillReportState bool                   `json:"willReportState"`
+	RoomHint        string                 `json:"roomHint,omitempty"`
+	Attributes      map[string]interface{} `json:"attributes,omitempty"`
+
+	DeviceInfo struct {
+		Manufacturer string `json:"manufacturer,omitempty"`
+		Model        string `json:"model,omitempty"`
+		HwVersion    string `json:"hwVersion,omitempty"`
+		SwVersion    string `json:"swVersion,omitempty"`
+	} `json:"deviceInfo,omitempty"`
+
+	OtherDeviceIDs []otherDeviceIDraw     `json:"otherDeviceIds,omitempty"`
+	CustomData     map[string]interface{} `json:"customData,omitempty"`
 }
