+++ go.mod
@@ -3,6 +3,7 @@ module github.com/rmrobinson/google-smart-home-action-go
 go 1.14
 
 require (
+	github.com/google/uuid v1.1.2
 	github.com/stretchr/testify v1.4.0
 	go.uber.org/zap v1.16.0
 	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
